Added parking code

+++ lib/pf/parking.pm
@@ -48,8 +48,20 @@ sub park {
     my ($mac,$ip) = @_;
     get_logger->debug("Setting client in parking");
     if(isenabled($Config{parking}{place_in_dhcp_parking_group})){
-        my $omapi = pf::OMAPI->get_client();
-        $omapi->create_host($mac, {group => $PARKING_DHCP_GROUP_NAME});
+        if (isenabled($Config{'services'}{'dhcpd'})){
+            my $omapi = pf::OMAPI->get_client();
+            $omapi->create_host($mac, {group => $PARKING_DHCP_GROUP_NAME});
+        } else {
+            pf::api::jsonrestclient->new(
+                proto   => "http",
+                host    => "localhost",
+                port    => $pf::constants::api::GO_DHCP_PORT,
+            )->call("/options/".$mac."/", [{
+                "option"      => "51",
+                "value"       => "3600",
+                "type"        => "int",
+            }]);
+        }
     }
     if(isenabled($Config{parking}{show_parking_portal})){
         my $cmd = "sudo ipset add $PARKING_IPSET_NAME $ip 2>&1";
@@ -86,10 +98,22 @@ Remove the parking actions that were taken against an IP + MAC
 sub remove_parking_actions {
     my ($mac, $ip) = @_;
     get_logger->info("Removing parking actions for $mac - $ip");
-    eval {
-        my $omapi = pf::OMAPI->get_client();
-        $omapi->delete_host($mac);
-    };
+    if (isenabled($Config{'services'}{'dhcpd'})){
+        eval {
+            my $omapi = pf::OMAPI->get_client();
+            $omapi->delete_host($mac);
+        };
+        if($@) {
+            get_logger->warn("Failed to remove client from parking using OMAPI ($@).");
+        }
+    } else {
+        pf::api::jsonrestclient->new(
+            proto   => "http",
+            host    => "localhost",
+            port    => $pf::constants::api::GO_DHCP_PORT,
+        )->call("/removeoptions/".$mac."/",
+        );
+    }
     if($@) {
         get_logger->warn("Failed to remove client from parking using OMAPI ($@).");
     }
